requirements.txt
opencv-python
opencv-contrib-python
numpy

RPi.GPIO
pigpio

smbus2
